Phase 1 Step 3: extract FundamentalsService
All 12 company_fundamentals tools, REST /earnings, and the watchlist
fundamentals report script now route through
quantcore.services.fundamentals.FundamentalsService. The MCP server is a
thin adapter (one service call per tool, docstrings verbatim).

- 8 yfinance access methods added to YFinanceGateway; info() is raw
  (no watchdog) to preserve original t.info blocking semantics
- FundamentalsRepository.ttl_seconds() exposes the cache TTL
- Pure helpers refactored to take DataFrames instead of yf.Ticker;
  math is verbatim
- Dead _qoq_vol_4 helper and a redundant quarterly-revenue fetch in
  score compute dropped (one fewer network call, identical output)
- api/app.py now imports get_services at module level
- Parity verified on 14 surfaces plus a fresh-compute check
